requirements.txt
fastapi
uvicorn[standard]
google-generativeai
faiss-cpu
numpy
python-dotenv
sentence-transformers
pydantic
